refactored configuretree
