+++ src/main/java/VASSAL/configure/ChooseComponentDialog.java
@@ -34,7 +34,7 @@ import javax.swing.tree.TreePath;
 import VASSAL.build.Buildable;
 import VASSAL.build.Configurable;
 import VASSAL.build.GameModule;
-import VASSAL.configure.ConfigureTree;
+import VASSAL.configure.configuretree.ConfigureTree;
 import VASSAL.tools.ScrollPane;
 
 /**
@@ -46,7 +46,7 @@ public class ChooseComponentDialog extends JDialog implements TreeSelectionListe
   private Configurable target;
   private Class<? extends Buildable> targetClass;
   private JButton okButton;
-  private VASSAL.configure.ConfigureTree tree;
+  private ConfigureTree tree;
 
   public ChooseComponentDialog(Frame owner, Class<? extends Buildable> targetClass) {
     super(owner, true);
