+++ src/main/java/VASSAL/tools/UniqueIdManager.java
@@ -22,7 +22,7 @@ import java.util.Iterator;
 import java.util.List;
 
 import VASSAL.build.Buildable;
-import VASSAL.configure.ConfigureTree;
+import VASSAL.configure.configuretree.ConfigureTree;
 import VASSAL.configure.ValidationReport;
 import VASSAL.configure.ValidityChecker;
